requirements.txt
streamlit==1.28.1
requests==2.31.0
folium==0.14.0
streamlit-folium==0.13.0
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0 